+++ tests/054_qkview.tests.ts
@@ -21,8 +21,8 @@ let testFile = '';
 let outFile = '';
 
 
-describe('qkview tests', async function() {
-    
+describe('qkview tests', async function () {
+
     before(async () => {
         testFile = await archiveMake('qkview') as string;
         const testFileDetails = path.parse(testFile);
@@ -31,15 +31,18 @@ describe('qkview tests', async function() {
         console.log('outFile', outFile);
     })
 
-    it(`instantiate -> parse configs, get parseTime, explode`, async function() {
-        
+    it(`instantiate -> parse configs, get parseTime, explode`, async function () {
+
         device = new BigipConfig();
 
-        device.on('parseFile', (x: any) => parsedFileEvents.push(x) )
-        device.on('parseObject', (x: any) => parsedObjEvents.push(x) )
+        device.on('parseFile', (x: any) => parsedFileEvents.push(x))
+        device.on('parseObject', (x: any) => parsedObjEvents.push(x))
 
-        const parseTime = await device.loadParseAsync(testFile);
-        expld = await device.explode();
+        const parseTime = await device.loadParseAsync(testFile)
+            .then(async parseTime => {
+                expld = await device.explode();
+                return parseTime;
+            });
 
         fs.writeFileSync(`${outFile}.json`, JSON.stringify(expld, undefined, 4));
         const bigLog = logOutput(device.configObject, expld);
@@ -49,7 +52,7 @@ describe('qkview tests', async function() {
 
     });
 
-    it(`list apps`, async function() {
+    it(`list apps`, async function () {
 
         const apps = await device.appList();
 
@@ -59,18 +62,20 @@ describe('qkview tests', async function() {
             "/Common/app2_t80_vs",
             "/Common/bigiq.benlab.io_t443_vs",
             "/Common/app2_t443_vs",
+            "/Common/persistTest_80_vs",
             "/Common/app3_t8443_vs",
             "/Common/app4_t80_vs",
             "/Common/forwarder_net_0.0.0.0",
             "/foo/defaultsUDP_5555/serviceMain",
+            "/foo/t1.lab.io_80vs",
             "/hue-infra/hue-up/hue-up.benlab.io_t80_vs",
             "/hue-infra/hue-up/hue-up.benlab.io_t443_vs",
           ]
-        
+
         assert.deepStrictEqual(apps, expected, 'Should get list of virtual servers / apps');
     });
 
-    it(`get app config by name`, async function() {
+    it(`get app config by name`, async function () {
 
         const app = await device.apps('/Common/app4_t80_vs');
         const expected = [
@@ -78,18 +83,18 @@ describe('qkview tests', async function() {
             "ltm pool /Common/app4_pool {\n    members {\n        /Common/api.chucknorris.io:443 {\n            fqdn {\n                autopopulate enabled\n                name api.chucknorris.io\n            }\n        }\n    }\n}",
             "ltm node /Common/api.chucknorris.io {\n    fqdn {\n        address-family all\n        autopopulate enabled\n        name api.chucknorris.io\n    }\n}",
             "ltm rule /Common/app4_pool_rule {\n### test rule for corkscrew\n\n  # \n\nwhen HTTP_REQUEST {\n\n  # pool reference by variable declaration\n  set html-pool web1Pool\n\n  if { [HTTP::path] ends_with \"*.css\" }{\n\n    # regular pool refernce\n    pool css_pool\n\n  } elseif { [HTTP::path] ends_with \"*.jpg\" }{\n\n    # pool member refernce\n    pool jpg.pool member 10.10.10.1 80\n\n  } elseif { [HTTP::path] ends_with \"*.js\" }{\n\n    # another pool reference with special characters\n    pool js.io_t80_pool \n\n  } elseif { [HTTP::path] ends_with \"*.xx\" }{\n\n    # pool reference not in tmos config\n    ### *** seems the gui won't let you attach an irule to a vs with a pool that doesn't exist\n    #pool missing_pool\n\n  } elseif { [HTTP::path] ends_with \"*.txt\" }{\n\n    # node reference\n    node 10.10.10.1 80\n\n  } else {\n\n    # pool referenced by variable\n    pool $html-pool\n\n  }\n}\n}",
-            "ltm pool /Common/css_pool { }",
-            "ltm pool /Common/jpg.pool { }",
-            "ltm pool /Common/js.io_t80_pool { }",
             "ltm policy /Common/app4_ltPolicy {\n    controls { forwarding }\n    description \"testing for pool extraction function\"\n    requires { http }\n    rules {\n        css_pool_rule {\n            actions {\n                0 {\n                    forward\n                    select\n                    pool /Common/css_pool\n                }\n            }\n            conditions {\n                0 {\n                    http-uri\n                    scheme\n                    ends-with\n                    values { .css }\n                }\n            }\n        }\n        jpg_pool_rule {\n            actions {\n                0 {\n                    forward\n                    select\n                    pool /Common/jpg.pool\n                }\n            }\n            conditions {\n                0 {\n                    http-uri\n                    query-string\n                    ends-with\n                    values { .jpg }\n                }\n            }\n            ordinal 1\n        }\n        js_pool_rule {\n            actions {\n                0 {\n                    forward\n                    select\n                    pool /Common/js.io_t80_pool\n                }\n            }\n            conditions {\n                0 {\n                    http-uri\n                    scheme\n                    ends-with\n                    values { .js }\n                }\n            }\n            ordinal 2\n        }\n        txt_node {\n            actions {\n                0 {\n                    forward\n                    select\n                    node 10.10.10.1\n                }\n            }\n            conditions {\n                0 {\n                    http-uri\n                    scheme\n                    ends-with\n                    values { .txt }\n                }\n            }\n            ordinal 3\n        }\n    }\n    strategy /Common/first-match\n}",
-          ];
+            "ltm pool css_pool { }",
+            "ltm pool jpg.pool { }",
+            "ltm pool js.io_t80_pool { }",
+          ]
 
-        const appConfig = app![0].config;
-        
+        const appConfig = app![0].lines;
+
         assert.deepStrictEqual(appConfig, expected, 'Should get list of virtual servers / apps');
     });
 
-    it(`only qvkiew SHOULD have default profiles/settings`, async function() {
+    it(`only qvkiew SHOULD have default profiles/settings`, async function () {
 
         const baseLtmProfiles = device.defaultProfileBase;
         const sysLowProfiles = device.defaultLowProfileBase;
@@ -99,15 +104,15 @@ describe('qkview tests', async function() {
     });
 
 
-    it(`basic gtm - high level fqdn details`, async function() {
+    it(`basic gtm - high level fqdn details`, async function () {
 
         const gslb = expld.config.gslb;
 
-        assert.ok(gslb!.length === 5);
-        assert.ok(typeof gslb![0] == 'object');
-        assert.ok(typeof gslb![0].fqdn == 'string');
-        assert.ok(typeof gslb![0].type == 'string');
-        assert.ok(typeof gslb![1].aliases![0] == 'string');
+        assert.ok(gslb!.length > 5);
+        assert.ok(typeof gslb![0] === 'object');
+        assert.ok(typeof gslb![0].fqdn === 'string');
+        assert.ok(typeof gslb![0].type === 'string');
+        assert.ok(typeof gslb![1].aliases![0] === 'string');
     });
 });