+++ CHANGELOG.md
@@ -7,6 +7,33 @@ All notable changes to the corkscrew rpm will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+- [RFE] include xml stats in extraction #15
+
+### Fixed
+
+- [BUG] no cli parse output in my environment #27
+
+---
+
+## [1.1.0] - (04.30.2023)
+
+### Added
+
+- [RFE] dns/gslb - provide all possible destinations as array #29
+- [RFE] better apm parsing #31
+- [RFE] deeper parsing of key ltm objects #30
+- [RFE] add bot/dos profiles to abstraction/visibility #33
+
+### Fixed
+
+- [BUG] gslb no pool error #32
+
 ---
 
 ## [1.0.0] - (04.20.2023)
@@ -86,7 +113,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - certificate/key extraction
 
-### Modified
+### Changed
 
 - Core parsing/extraction process is now asyncronous
   - emits config and stats files during processing